@@ -15,12 +15,20 @@ namespace ECS {
             Movable();
             ~Movable();
 
-            void moveUp();
-            void moveDown();
-            void moveLeft();
-            void moveRight();
         protected:
+            using ECS::Drawable::setSprite;
+            using ECS::Object::getId;
+            using ECS::Object::getPosX;
+            using ECS::Object::getPosY;
+            using ECS::Object::getType;
+            using ECS::Object::setId;
+            using ECS::Object::setPosY;
+            using ECS::Object::setPosX;
 
+            // Utile ???
+            using ECS::Object::setType;
+            using ECS::Object::isEmpty;
+            using ECS::Object::setEmpty;
         private:
     };
 }
