@@ -7,12 +7,22 @@
 
 #pragma once
 
+#include "Movable.hpp"
+
 namespace ECS {
-    class MovableEntity {
+    enum Direction {
+        UP,
+        DOWN,
+        LEFT,
+        RIGHT
+    };
+    class MovableEntity : public ECS::Movable {
         public:
             MovableEntity();
             ~MovableEntity();
 
+            bool canMove(enum Direction direction, std::vector<std::string> map);
+            bool move(enum Direction direction, std::vector<std::string> map);
         protected:
         private:
     };
